Allow nooped plurals to receive a textdomain on registration to then be used on translation.
This is good for when the code registering the plural is not also translating it. This occurs
in core with register_post_status(), which accepts a nooped plural as an argument, and then
calls translate_nooped_plural() without a domain.

translate_nooped_plural() can still be called with a domain. The argument will just be overridden
if the nooped plural contains a domain key.

fixes #20188.



git-svn-id: https://develop.svn.wordpress.org@20648 602fd350-edb4-49c9-b593-d223f7449a82