Introduce a `.dashicons-before` CSS class.
If you want to use a Dashicon before an element you can use the class because it will not change the elements content styling. With that you don't need to copy the entire `.dashicons` styling to your :before styling anymore.
To demonstrate this change, Admin Menu now uses Dashicons classes directly.

And it fixes a glitch in IE.

fixes #26630.

git-svn-id: https://develop.svn.wordpress.org@27418 602fd350-edb4-49c9-b593-d223f7449a82