In `wp_import_handle_upload()`:
`$file` was essentially getting declared/overwritten 3 times. In lieu of this, return an array containing the error immediately instead of doing a short-circuit array key assignment on error. Rename the local variable to `$upload` and use its properties instead of creating 3 new local vars, one of which stomped the array.

See #30799.


git-svn-id: https://develop.svn.wordpress.org@30980 602fd350-edb4-49c9-b593-d223f7449a82