List tables: Yet more primary column fallbacks.
Some custom list tables override enough methods for the column definition fallback to never kick in, so let's ensure that toggling columns only applies when a primary column is defined in some way. We also need to show a toggle button when we can when there are no row actions.

props Chouby, obenland, ocean90.
fixes #33313.


git-svn-id: https://develop.svn.wordpress.org@33623 602fd350-edb4-49c9-b593-d223f7449a82