Pass the `$public_only` value to 'get_usernumposts' filter.
[32523] introduced the $public_only parameter to `count_user_posts()`. That
changeset was supposed to pass `$public_only` to the 'get_usernumposts' filter
at the end of the function, but only the documentation was modified, not the
filter itself.

This changeset also fixes an incorrect variable name in the docblock for
the same filter, and includes clarification on the accepted values for
the `$post_type` parameter.

Merges [33710] and [33716] to the 4.3 branch.

Props swisspidy, tmatsuur, tyxla, DrewAPicture.
Fixes #33481 for 4.3.1. 


git-svn-id: https://develop.svn.wordpress.org@33953 602fd350-edb4-49c9-b593-d223f7449a82