Revert change from true to false in multisite upload test
This change was introduced in [34601]. 0 is a special snowflake. The code treats it as if it should fall back to the default, rather than respecting 0 as 0.  Essentially, 0 means 100.  WAT? This has been discussed previusly in https://core.trac.wordpress.org/ticket/19538#comment:10.  This test should return True right now, but the behavior should be discussed and in the future 0 may mean 0.

See #34037
Props jorbin, boonebgorges



git-svn-id: https://develop.svn.wordpress.org@34606 602fd350-edb4-49c9-b593-d223f7449a82