Customize: Fix sub-nav menu items from appearing to bump to root level in preview upon saving.
This issue would appear when the parent nav menu item was also newly-inserted. The issue was only apparent in the prevew, as the actual data saved was correct. The `menu_item_parent` properties of the saved sub-nav menu items needed to get updated to refer to the newly-inserted parent `nav_menu_item` post IDs, as opposed to retaining the placeholder IDs.

Fixes #34628.


git-svn-id: https://develop.svn.wordpress.org@35581 602fd350-edb4-49c9-b593-d223f7449a82