Template: Defining a default value for `show_home` breaks back compat.
To add a home link to the fallback menu output many themes only check if that
argument is set. Including Twenty Ten and Twenty Eleven. They check with
`isset()` so child themes and other instances using `wp_page_menu()` have a
chance to disable the home link by setting it to `false`.

Fixes #11095.



git-svn-id: https://develop.svn.wordpress.org@35737 602fd350-edb4-49c9-b593-d223f7449a82