I18N: Don't use the locale for the HTML language attribute.
Locales are codes to identify a language in WordPress which can be different from the specification for language tags, see https://www.w3.org/International/articles/language-tags/.
An example is `de_DE_formal` or `nl_NL_formal` where the subtag `formal` isn't officially supported.

To give translators the possibility to specify the language tag of their language introduce a string `html_lang_attribute` which can be translated into the language tag which conforms to the specification.

Props SergeyBiryukov.
Fixes #33511.

git-svn-id: https://develop.svn.wordpress.org@36802 602fd350-edb4-49c9-b593-d223f7449a82