Customize: Rely on selective refresh exclusively for previewing custom logo changes.
Eliminates JS logic (from [36698]) which attempted to do pure JS update while waiting for the selective refresh response to return. The duplicate JS logic lacked a re-implementation of the `image_downsize()` functionality available in PHP, and so the JS preview logic would fail to properly preview images that didn't have the exact theme image size generated. To keep the code DRY and to eliminate the momentary display of an improperly-sized image, the duplicated JS logic is now removed.

See #27355.
See #33755.
Fixes #36096.


git-svn-id: https://develop.svn.wordpress.org@36990 602fd350-edb4-49c9-b593-d223f7449a82