Use `px` instead of `in` in device preview
Devices are not consistent in how they handle `in` units. `in` was an attempt to cleverly disguise the exact size of the 'tablet'. The PHP code standards mentions avoiding clever code ( https://make.wordpress.org/core/handbook/best-practices/coding-standards/php/#clever-code ) but we should extend that idea to the css code as well.

Props celloexpressions
Fixes #36457



git-svn-id: https://develop.svn.wordpress.org@37247 602fd350-edb4-49c9-b593-d223f7449a82