Media: Remove medium_large size from `$whitelist_options['media']` in `options.php`.
The `medium_large_size_w` and `medium_large_size_h` options were added to the
`$whitelist_options['media']` in options.php in [35479], which causes both options
to be set to `0` when the media settings are changed. When this occurs, the
medium_large image size will no longer be created when an image is uploaded.

Since these options aren't meant to be updated via the settings UI, they can be
safely removed from the whitelist.

Merge of [37435] to the 4.5 branch.

Props joemcgill.
Fixes #36531.

git-svn-id: https://develop.svn.wordpress.org@37457 602fd350-edb4-49c9-b593-d223f7449a82