Updates: Improve experience for Bulk Actions when FTP is dismissed.
Merges [38221] to the 4.6 branch.

Before this change, when a bulk update was canceled due dismissing the FTP credentials modal, part of the actions didn't get canceled. This meant the "There is a new version of…” notices become blank and the updates you had checked became unchecked. Now, the notices remain and you are essentially returned to the screen you had before. Strings are also updated to improve ARIA usage.

Fixes #37563.
Props ocean90, swissspidy, obenland, afercia.



git-svn-id: https://develop.svn.wordpress.org@38222 602fd350-edb4-49c9-b593-d223f7449a82